+++ source/redux/translate/reducer.js
@@ -1,0 +1,4 @@
+// Not implemented.
+export default function(state = {}, event) {
+	return state
+}